requirements.txt
streamlit
streamlit-autorefresh
pandas
requests
pyserial
pymodbus
fpdf2
apscheduler
fastapi
uvicorn
